jv - clean up login/register ui + wip navigation
