--- app/src/main/java/edu/ucsb/cs/cs184/group9/billsplitter/ui/login/LoginPage.kt
@@ -1,235 +1,0 @@
-package edu.ucsb.cs.cs184.group9.billsplitter.ui.login
-
-import android.util.Log
-import androidx.activity.ComponentActivity
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-//import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.runtime.*
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.imageResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
-//import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.TextUnit
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
-//import androidx.navigation.compose.navigate
-import edu.ucsb.cs.cs184.group9.billsplitter.ui.theme.primaryColor
-import edu.ucsb.cs.cs184.group9.billsplitter.ui.theme.whiteBackground
-
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.Text
-import androidx.compose.material.TextField
-import androidx.compose.material.icons.filled.Visibility
-import androidx.compose.material.icons.filled.VisibilityOff
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.navigation.compose.rememberNavController
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-import edu.ucsb.cs.cs184.group9.billsplitter.ui.nav.Screen
-
-
-//@Preview
-@Composable
-fun LoginPage(context: ComponentActivity) {
-    val navController = rememberNavController()
-
-    val auth = Firebase.auth
-    Column(
-        modifier = Modifier
-            .fillMaxSize(),
-            verticalArrangement = Arrangement.SpaceEvenly,
-            horizontalAlignment = Alignment.CenterHorizontally
-    ){
-        Text(
-            text = "Login",
-            fontSize = 64.sp,
-            fontWeight = FontWeight.Bold,
-            color = primaryColor,
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .padding(10.dp)
-                .fillMaxWidth()
-        )
-
-        var email by remember { mutableStateOf("") }
-        TextField(
-            value = email,
-            onValueChange = { email = it },
-            singleLine = true,
-            label = { Text("Email") }
-        )
-
-        var password by rememberSaveable { mutableStateOf("") }
-        var passwordHidden by rememberSaveable { mutableStateOf(true) }
-        TextField(
-            value = password,
-            onValueChange = { password = it },
-            singleLine = true,
-            label = { Text("Password") },
-            visualTransformation =
-            if (passwordHidden) PasswordVisualTransformation() else VisualTransformation.None,
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-            trailingIcon = {
-                IconButton(onClick = { passwordHidden = !passwordHidden }) {
-                    val visibilityIcon =
-                        if (passwordHidden) Icons.Filled.Visibility else Icons.Filled.VisibilityOff
-                    // Please provide localized description for accessibility services
-                    val description = if (passwordHidden) "Show password" else "Hide password"
-                    Icon(imageVector = visibilityIcon, contentDescription = description)
-                }
-            }
-        )
-
-
-        Button(onClick = {
-            auth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(context) { task ->
-                    if (task.isSuccessful) {
-                        // Sign in success, update UI with the signed-in user's information
-                        Log.d("", "signInWithEmail:success")
-                        val user = auth.currentUser
-//                        updateUI(user)
-                    } else {
-                        // If sign in fails, display a message to the user.
-                        Log.w("", "signInWithEmail:failure", task.exception)
-                    }
-                }
-        })
-        {
-            Text("Login")
-        }
-
-        Row(
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically
-
-        ){
-            Text(
-                text = "Not registered yet? ",
-                fontSize = 16.sp,
-            )
-            TextButton(onClick = {
-                navController.navigate(route = Screen.Register.route)
-            })
-            {
-                Text("Register here")
-            }
-        }
-
-    }
-}
-
-@Composable
-fun RegisterPage(context: ComponentActivity) {
-    val navController = rememberNavController()
-
-    val auth = Firebase.auth
-    Column(
-        modifier = Modifier
-            .fillMaxSize(),
-        verticalArrangement = Arrangement.SpaceEvenly,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ){
-        Text(
-            text = "Register",
-            fontSize = 64.sp,
-            fontWeight = FontWeight.Bold,
-            color = primaryColor,
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .padding(10.dp)
-                .fillMaxWidth()
-        )
-
-        var email by remember { mutableStateOf("") }
-        TextField(
-            value = email,
-            onValueChange = { email = it },
-            singleLine = true,
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
-            label = { Text("Email") }
-        )
-
-        var password by rememberSaveable { mutableStateOf("") }
-        var passwordHidden by rememberSaveable { mutableStateOf(true) }
-        TextField(
-            value = password,
-            onValueChange = { password = it },
-            singleLine = true,
-            label = { Text("Password") },
-            visualTransformation =
-            if (passwordHidden) PasswordVisualTransformation() else VisualTransformation.None,
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-            trailingIcon = {
-                IconButton(onClick = { passwordHidden = !passwordHidden }) {
-                    val visibilityIcon =
-                        if (passwordHidden) Icons.Filled.Visibility else Icons.Filled.VisibilityOff
-                    // Please provide localized description for accessibility services
-                    val description = if (passwordHidden) "Show password" else "Hide password"
-                    Icon(imageVector = visibilityIcon, contentDescription = description)
-                }
-            }
-        )
-
-
-        Button(onClick = {
-            auth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(context) { task ->
-                    if (task.isSuccessful) {
-                        // Sign in success, update UI with the signed-in user's information
-                        Log.d("Auth", "createUserWithEmail:success")
-                        val user = auth.currentUser
-//                        MainScreen()
-                    } else {
-                        // If sign in fails, display a message to the user.
-                        Log.w("Auth", "createUserWithEmail:failure", task.exception)
-                    }
-                }
-
-        })
-        {
-            Text("Register")
-        }
-
-        Row(
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically
-        ){
-            Text(
-                text = "Already registered? ",
-                fontSize = 16.sp,
-            )
-            TextButton(onClick = {
-                navController.navigate(route = Screen.Login.route)
-            })
-            {
-                Text("Login here")
-            }
-        }
-
-    }
-}
